Add OSX version and restructre project
- Add files from the OSX version and its depencency to WledLib
- Split directory `wled-native-iOS`:
    - Shared files:        `wled-native`
    - Exclusive iOS files: `wled-native-iOS`
    - Exclusuve OSX files: `wled-native-osx`
- Resolve setup issues from move of files
- Extend `README.md` with the OSX content
