+++ wled-native-macOS/Assets.xcassets/012_rip.imageset/012_rip.png
