+++ wled-native-macOS/Assets.xcassets/012_rip.imageset/Contents.json
