+++ wled-native-macOS/Assets.xcassets/AccentColor.colorset/Contents.json
