+++ wled-native-macOS/Assets.xcassets/AppIconMac.appiconset/Contents.json
