+++ wled-native-macOS/Assets.xcassets/AppIconMac.appiconset/mac-128x128.png
