+++ wled-native-macOS/Assets.xcassets/AppIconMac.appiconset/mac-16x16@2x.png
