+++ wled-native-macOS/Assets.xcassets/AppIconMac.appiconset/mac-256x256.png
