+++ wled-native-macOS/Assets.xcassets/AppIconMac.appiconset/mac-32x32.png
