+++ wled-native-macOS/Assets.xcassets/AppIconMac.appiconset/mac-512x512.png
