+++ wled-native-macOS/Assets.xcassets/Contents.json
