+++ wled-native-macOS/Assets.xcassets/wled_logo_akemi.imageset/Contents.json
