+++ wled-native-macOS/Assets.xcassets/wled_logo_akemi.imageset/wled_logo_akemi.png
