+++ wled-native-macOS/Documentation.docc/Collections/View.md
@@ -1,6 +1,6 @@
 # View
 
-The Views of the 
+The Views of the macOS application.
 
 ## Overview
