+++ wled-native-macOS/View/DeviceView.swift
@@ -1,6 +1,6 @@
 //
 //  DeviceView.swift
-//  wled-osx
+//  wled-native-macOS
 //
 //  Created by Robert Brune on 15.11.23.
 //
