+++ wled-native-macOS/View/Main.swift
@@ -1,6 +1,6 @@
 //
 //  Main.swift
-//  wled-osx
+//  wled-native-macOS
 //
 //  Created by Robert Brune on 15.11.23.
 //
