+++ wled-native-macOS/ViewProtocolExtensions/DeviceProviderExtension.swift
@@ -1,6 +1,6 @@
 //
 //  DeviceProviderExtension.swift
-//  wled-osx
+//  wled-native-macOS
 //
 //  Created by Robert Brune on 26.11.23.
 //
