+++ wled-native-macOS/ViewProtocolExtensions/WledDeviceExtension.swift
@@ -1,6 +1,6 @@
 //
 //  WledDeviceExtension.swift
-//  wled-osx
+//  wled-native-macOS
 //
 //  Created by Robert Brune on 26.11.23.
 //
