+++ wled-native-macOS/ViewProtocolls/CollectionProvider.swift
@@ -1,6 +1,6 @@
 //
 //  CollectionProvider.swift
-//  wled-osx
+//  wled-native-macOS
 //
 //  Created by Robert Brune on 26.11.23.
 //
