+++ wled-native-macOS/wled-macOS.entitlements
