+++ wled-native-macOS/wled-native-macOS.md
@@ -1,0 +1,6 @@
+#  wled native macOS
+
+This directory should only contain files used by the wled native macOS app.
+
+
+
