--- wled-native-osx/wled-native-osx.md
@@ -1,6 +1,0 @@
-#  wled native osx
-
-This directory should only contain files used by the wled native OSX app.
-
-
-
