+++ wled-native.xcodeproj/xcshareddata/xcschemes/wled-native.xcscheme
@@ -15,8 +15,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "C9A9E5BC2918D1C3005B5C22"
-               BuildableName = "wled-native-ios.app"
-               BlueprintName = "wled-native-ios"
+               BuildableName = "wled-native-iOS.app"
+               BlueprintName = "wled-native-iOS"
                ReferencedContainer = "container:wled-native.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -44,8 +44,8 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "C9A9E5BC2918D1C3005B5C22"
-            BuildableName = "wled-native-ios.app"
-            BlueprintName = "wled-native-ios"
+            BuildableName = "wled-native-iOS.app"
+            BlueprintName = "wled-native-iOS"
             ReferencedContainer = "container:wled-native.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
@@ -61,8 +61,8 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "C9A9E5BC2918D1C3005B5C22"
-            BuildableName = "wled-native-ios.app"
-            BlueprintName = "wled-native-ios"
+            BuildableName = "wled-native-iOS.app"
+            BlueprintName = "wled-native-iOS"
             ReferencedContainer = "container:wled-native.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
